+++ lib/zip_fopen_index_encrypted.c
@@ -49,7 +49,7 @@ zip_fopen_index_encrypted(zip_t *za, zip_uint64_t index, zip_flags_t flags, cons
         password = NULL;
     }
     
-    if ((src = _zip_source_zip_new(za, index, flags, 0, 0, password, &za->error)) == NULL)
+    if ((src = zip_source_zip_file_create(za, index, flags, 0, -1, password, &za->error)) == NULL)
         return NULL;
 
     if (zip_source_open(src) < 0) {