+++ src/ziptool.c
@@ -109,7 +109,7 @@ cat_impl(zip_uint64_t idx, zip_uint64_t start, zip_uint64_t len) {
         }
         len = sb.size;
     }
-    if ((src = zip_source_zip_create(za, idx, 0, start, len, &error)) == NULL) {
+    if ((src = zip_source_zip_file_create(za, idx, 0, start, len, NULL, &error)) == NULL) {
         fprintf(stderr, "can't open file at index '%" PRIu64 "': %s\n", idx, zip_error_strerror(&error));
         zip_error_fini(&error);
         return -1;
@@ -152,7 +152,7 @@ add(char *argv[]) {
         return -1;
     }
 
-    if (zip_add(za, argv[0], zs) == -1) {
+    if (zip_file_add(za, argv[0], zs, 0) == -1) {
         zip_source_free(zs);
         fprintf(stderr, "can't add file '%s': %s\n", argv[0], zip_strerror(za));
         return -1;
@@ -163,7 +163,7 @@ add(char *argv[]) {
 static int
 add_dir(char *argv[]) {
     /* add directory */
-    if (zip_add_dir(za, argv[0]) < 0) {
+    if (zip_dir_add(za, argv[0], 0) < 0) {
         fprintf(stderr, "can't add directory '%s': %s\n", argv[0], zip_strerror(za));
         return -1;
     }
@@ -189,7 +189,7 @@ add_file(char *argv[]) {
         }
     }
 
-    if (zip_add(za, argv[0], zs) == -1) {
+    if (zip_file_add(za, argv[0], zs, 0) == -1) {
         zip_source_free(zs);
         fprintf(stderr, "can't add file '%s': %s\n", argv[0], zip_strerror(za));
         return -1;
@@ -203,6 +203,7 @@ add_from_zip(char *argv[]) {
     zip_int64_t len;
     int err;
     zip_source_t *zs;
+    zip_flags_t flags = 0;
     /* add from another zip file */
     idx = strtoull(argv[2], NULL, 10);
     start = strtoull(argv[3], NULL, 10);
@@ -214,12 +215,15 @@ add_from_zip(char *argv[]) {
         zip_error_fini(&error);
         return -1;
     }
-    if ((zs = zip_source_zip(za, z_in[z_in_count], idx, 0, start, len)) == NULL) {
+    if (start == 0 && len == -1) {
+        flags = ZIP_FL_COMPRESSED;
+    }
+    if ((zs = zip_source_zip_file(za, z_in[z_in_count], idx, flags, start, len, NULL)) == NULL) {
         fprintf(stderr, "error creating file source from '%s' index '%" PRIu64 "': %s\n", argv[1], idx, zip_strerror(za));
         zip_close(z_in[z_in_count]);
         return -1;
     }
-    if (zip_add(za, argv[0], zs) == -1) {
+    if (zip_file_add(za, argv[0], zs, 0) == -1) {
         fprintf(stderr, "can't add file '%s': %s\n", argv[0], zip_strerror(za));
         zip_source_free(zs);
         zip_close(z_in[z_in_count]);
@@ -402,18 +406,18 @@ get_extra_by_id(char *argv[]) {
 static int
 get_file_comment(char *argv[]) {
     const char *comment;
-    int len;
+    zip_uint32_t len;
     zip_uint64_t idx;
     /* get file comment */
     idx = strtoull(argv[0], NULL, 10);
-    if ((comment = zip_get_file_comment(za, idx, &len, 0)) == NULL) {
+    if ((comment = zip_file_get_comment(za, idx, &len, 0)) == NULL) {
         fprintf(stderr, "can't get comment for '%s': %s\n", zip_get_name(za, idx, 0), zip_strerror(za));
         return -1;
     }
     else if (len == 0)
         printf("No comment for '%s'\n", zip_get_name(za, idx, 0));
     else
-        printf("File comment for '%s': %.*s\n", zip_get_name(za, idx, 0), len, comment);
+        printf("File comment for '%s': %.*s\n", zip_get_name(za, idx, 0), (int)len, comment);
     return 0;
 }
 
@@ -467,7 +471,7 @@ static int
 zrename(char *argv[]) {
     zip_uint64_t idx;
     idx = strtoull(argv[0], NULL, 10);
-    if (zip_rename(za, idx, argv[1]) < 0) {
+    if (zip_file_rename(za, idx, argv[1], 0) < 0) {
         fprintf(stderr, "can't rename file at index '%" PRIu64 "' to '%s': %s\n", idx, argv[1], zip_strerror(za));
         return -1;
     }