+++ NEWS.md
@@ -1,6 +1,7 @@
-# 1.10 [Unreleased]
+# 1.10.0 [Unreleased]
 
 * Make support for layered sources public.
+* Add `zip_source_zip_file` and `zip_source_zip_file_create`, deprecate `zip_source_zip` and `zip_source_zip_create`.
 * Allow reading changed file data.
 * Fix handling of files of size 4294967295.
 * `zipmerge`: copy extra fields.
@@ -9,6 +10,7 @@
 * Use ISO C secure library functions, if available.
 * Fix reading/writing compressed data with buffers > 4GiB.
 * Restore support for torrentzip.
+* Add warnings when using deprecated functions.
 
 
 # 1.9.2 [2022-06-28]